api-service/requirements.txt
fastapi==0.110.1
ulid-py==1.1.0
jinja2==3.1.3
uvicorn==0.29.0
python-multipart==0.0.9
boto3==1.34.103
python-dotenv==1.0.1
